+++ src/material.rs
@@ -31,11 +31,11 @@ impl Material for Lambertian {
 #[derive(Debug)]
 pub struct Metal {
     albedo: Color,
-    fuzz: f32,
+    fuzz: f64,
 }
 
 impl Metal {
-    pub fn new(albedo: Color, fuzziness: f32) -> Self {
+    pub fn new(albedo: Color, fuzziness: f64) -> Self {
         let fuzz = if fuzziness < 1.0 { fuzziness } else { 1.0 };
         Metal { albedo, fuzz }
     }
@@ -55,11 +55,11 @@ impl Material for Metal {
 
 #[derive(Debug)]
 pub struct Dielectric {
-    pub ref_idx: f32,
+    pub ref_idx: f64,
 }
 
 impl Dielectric {
-    pub fn new(ref_idx: f32) -> Self {
+    pub fn new(ref_idx: f64) -> Self {
         Dielectric { ref_idx }
     }
 }
@@ -73,11 +73,11 @@ impl Material for Dielectric {
             (rec.normal, 1.0 / self.ref_idx)
         };
     
-        let cos_theta = f32::min(Vec3::dot(-r_in.dir.unit_along(), outward_normal), 1.0);
+        let cos_theta = f64::min(Vec3::dot(-r_in.dir.unit_along(), outward_normal), 1.0);
         let sin_theta = (1.0 - cos_theta.powi(2)).sqrt();
         let cannot_refract = ni_by_nt * sin_theta > 1.0;
         let mut rng = thread_rng();
-        let val: f32 = rng.gen();
+        let val: f64 = rng.gen();
 
         let direction = if cannot_refract || reflectance(cos_theta, ni_by_nt) > val {
             Vec3::reflect(r_in.dir.unit_along(), outward_normal)
@@ -93,7 +93,7 @@ impl Material for Dielectric {
     }
 }
 
-fn reflectance(cosine: f32, ref_idx: f32) -> f32 {
+fn reflectance(cosine: f64, ref_idx: f64) -> f64 {
     let r0 = ((1.0 - ref_idx) / (1.0 + ref_idx)).powi(2);
     r0 + (1.0 - r0) * (1.0 - cosine).powi(5)
 }
